separate `login and register` page and integrate rkt mutations
