+++ backend/controller/userController.js
@@ -85,7 +85,7 @@ const signup = async (req, res) => {
     // look up the user collection to see user have already exist or not.
     const userList = await getUsersByUsername(req.body.username);
     if (userList.length > 0) {
-      return res.status(200).json({
+      return res.status(409).json({
         message: "Username already exist!",
       });
     }
