--- frontend/src/store/actions/actionTypes.js
@@ -1,22 +1,0 @@
-// Auth
-export const AUTH_START = "AUTH_START";
-export const AUTH_SUCCESS = "AUTH_SUCCESS";
-export const AUTH_FAIL = "AUTH_FAIL";
-export const LOGOUT = "LOGOUT";
-
-// vault
-export const VAULT_START = "VAULT_START";
-export const VAULT_SUCCESS = "VAULT_SUCCESS";
-export const VAULT_FAIL = "VAULT_FAIL";
-
-export const ALL_VAULT_SUCCESS = "ALL_VAULT_SUCCESS";
-export const TEMP_VAULT_SUCCESS = "TEMP_VAULT_SUCCESS";
-
-//profile
-export const PROFILE_START = "PROFILE_START";
-export const PROFILE_SUCCESS = "PROFILE_SUCCESS";
-export const PROFILE_FAIL = "PROFILE_FAIL";
-
-// error modal
-export const OPEN_ERROR_MODAL = "OPEN_ERROR_MODAL";
-export const CLOSE_ERROR_MODAL = "CLOSE_ERROR_MODAL";
