--- frontend/src/store/actions/errorModal/errorModal.js
@@ -1,17 +1,0 @@
-import * as actionTypes from "../actionTypes";
-
-export const openErrorModal = (data) => {
-  return (dispatch) => {
-    dispatch({
-      type: actionTypes.OPEN_ERROR_MODAL,
-      data,
-    });
-  };
-};
-export const closeErrorModal = () => {
-  return async (dispatch) => {
-    dispatch({
-      type: actionTypes.CLOSE_ERROR_MODAL,
-    });
-  };
-};
