--- frontend/src/store/actions/index.js
@@ -1,9 +1,0 @@
-export { singIn, singUp } from "./auth/auth";
-export {
-  addCredential,
-  updateCredential,
-  removeCredential,
-  getAllCredentials,
-} from "./vault/vault";
-export { fetchProfile, updateProfile } from "./profile/profile";
-export { openErrorModal, closeErrorModal } from "./errorModal/errorModal";
